Move binary version variables to a seperate package
